# docs/swagger.yaml
basePath: https://auth.shariski.com
definitions:
  internal_handler.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  internal_handler.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/internal_handler.ErrorBody'
    type: object
  internal_handler.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_handler.LogoutRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  internal_handler.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  internal_handler.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  internal_handler.TokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  internal_handler.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      role:
        type: string
    type: object
  internal_handler.UsersListResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/internal_handler.UserResponse'
        type: array
    type: object
info:
  contact:
    email: falahudin6@gmail.com
    name: Falahudin Halim Shariski
  description: Auth API for backend infra assessment
  title: Auth API
  version: "1.0"
paths:
  /admin/users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.UsersListResponse'
        "401":
          description: missing or invalid access token
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "403":
          description: caller is not Admin
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "429":
          description: rate-limited
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all users (admin)
      tags:
      - admin
  /auth/login:
    post:
      consumes:
      - application/json
      description: Verifies credentials and returns JWT access and refresh tokens.
        Subject to per-IP brute-force protection and per-account lockout.
      parameters:
      - description: User credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.TokenResponse'
        "400":
          description: validation error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: invalid credentials
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "429":
          description: account locked or rate-limited
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Log in
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Revokes the refresh token. The access token remains valid until
        it expires.
      parameters:
      - description: Refresh token to revoke
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.LogoutRequest'
      produces:
      - application/json
      responses:
        "204":
          description: no content
        "400":
          description: validation error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: missing or invalid access token
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Log out
      tags:
      - auth
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.UserResponse'
        "401":
          description: missing or invalid access token
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Current user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Rotates the refresh token. The previous refresh token is invalidated
        on success.
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.TokenResponse'
        "400":
          description: validation error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: refresh token invalid, revoked, or expired
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a Viewer-role account. Returns the created user (without
        the password hash).
      parameters:
      - description: Email and password (min 8 chars)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.UserResponse'
        "400":
          description: validation error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "409":
          description: email already taken
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: 'Paste the access_token from /auth/login as: "Bearer <token>".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"