# internal/handler/dto.go
package handler

import "auth/internal/domain"

// Request DTOs — validation tags are enforced by Gin's binding.

type RegisterRequest struct {
	Email    string `json:"email" binding:"required,email"`
	Password string `json:"password" binding:"required,min=8"`
}

type LoginRequest struct {
	Email    string `json:"email" binding:"required,email"`
	Password string `json:"password" binding:"required"`
}

type RefreshRequest struct {
	RefreshToken string `json:"refresh_token" binding:"required"`
}

type LogoutRequest struct {
	RefreshToken string `json:"refresh_token" binding:"required"`
}

// Response DTOs.

type UserResponse struct {
	ID    string `json:"id"`
	Email string `json:"email"`
	Role  string `json:"role"`
}

type TokenResponse struct {
	AccessToken  string `json:"access_token"`
	RefreshToken string `json:"refresh_token"`
}

type UsersListResponse struct {
	Users []UserResponse `json:"users"`
}

func toUserResponse(u *domain.User) UserResponse {
	return UserResponse{
		ID:    u.ID.String(),
		Email: u.Email,
		Role:  string(u.Role),
	}
}